Moving info around. Making sure things are user related.

+++ Fuyuki.Server/Services/GroupService.cs
@@ -1,5 +1,7 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Fuyuki.Data;
@@ -9,36 +11,49 @@ namespace Fuyuki.Services
 {
     public class GroupService : IGroupService
     {
+        private readonly IUserService _userService;
         private readonly IGroupDataService _groupDataService;
         private readonly IMapper _mapper;
 
-        public GroupService(IGroupDataService groupDataService, IMapper mapper)
+        public GroupService(IUserService userService, IGroupDataService groupDataService, IMapper mapper)
         {
+            _userService = userService;
             _groupDataService = groupDataService;
             _mapper = mapper;
         }
 
-        public async Task<GroupSubsModel> GetGroupById(int id)
+        public async Task<GroupSubsModel> GetGroupById(ClaimsPrincipal claim, int id)
         {
+            var user = await _userService.GetUserByName(claim.Identity.Name);
             var group = await _groupDataService.GetAsync<Group>(id);
+
+            if (group.ApplicationUserId != user.Id)
+            {
+                throw new Exception("User is not authourised to view this group.");
+            }
+
             return _mapper.Map<GroupSubsModel>(group);
         }
 
-        public async Task<List<GroupModel>> GetGroups()
+        public async Task<List<GroupModel>> GetGroups(ClaimsPrincipal claim)
         {
-            var groups = await _groupDataService.GetGroups();
+            var user = await _userService.GetUserByName(claim.Identity.Name);
+            var groups = await _groupDataService.GetGroups(user.Id);
             return _mapper.Map<List<GroupModel>>(groups);
         }
 
-        public async Task<List<GroupSubsModel>> GetGroupsWithSubreddit()
+        public async Task<List<GroupSubsModel>> GetGroupsWithSubreddit(ClaimsPrincipal claim)
         {
-            var groups = await _groupDataService.GetGroups();
+            var user = await _userService.GetUserByName(claim.Identity.Name);
+            var groups = await _groupDataService.GetGroupsWithSubreddits(user.Id);
             return _mapper.Map<List<GroupSubsModel>>(groups);
         }
 
-        public async Task<GroupResponse> CreateGroup(GroupRequest request)
+        public async Task<GroupResponse> CreateGroup(ClaimsPrincipal claim, GroupRequest request)
         {
+            var user = await _userService.GetUserByName(claim.Identity.Name);
             var group = _mapper.Map<Group>(request);
+            group.ApplicationUserId = user.Id;
 
             foreach (var sub in request.Subreddits)
             {
@@ -58,9 +73,18 @@ namespace Fuyuki.Services
             };
         }
 
-        public async Task<GroupResponse> UpdateGroup(GroupRequest request)
+        public async Task<GroupResponse> UpdateGroup(ClaimsPrincipal claim, GroupRequest request)
         {
+            var user = await _userService.GetUserByName(claim.Identity.Name);
             var group = await _groupDataService.GetAsync<Group>(request.Id);
+
+            if (group.ApplicationUserId != user.Id)
+            {
+                return new GroupResponse
+                {
+                    ErrorMessages = new List<string> { "The group selected is not associated to the current user." }
+                };
+            }
 
             _mapper.Map(request, group);
 
@@ -86,9 +110,18 @@ namespace Fuyuki.Services
             };
         }
 
-        public async Task<GroupResponse> DeleteGroup(int id)
+        public async Task<GroupResponse> DeleteGroup(ClaimsPrincipal claim, int id)
         {
+            var user = await _userService.GetUserByName(claim.Identity.Name);
             var group = await _groupDataService.GetAsync<Group>(id, x => x.GroupSubreddits);
+
+            if (group.ApplicationUserId != user.Id)
+            {
+                return new GroupResponse
+                {
+                    ErrorMessages = new List<string> { "The group selected is not associated to the current user and could not be deleted." }
+                };
+            }
 
             _groupDataService.Delete(group);
